// admin/src/api/generated/schemas/adminVaultOverviewSummary.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */

export type AdminVaultOverviewSummary = {
  total_balance_usd?: string;
  total_networks?: number;
  healthy_networks?: number;
  pending_adjustments?: number;
  alerts_count?: number;
  unswept_balance_usd?: string;
};